+++ Plugin.cs
@@ -28,7 +28,7 @@ namespace HeavyItemSCPs
     {
         const string PLUGIN_GUID = "ProjectSCP.HeavyItemSCPs";
         const string PLUGIN_NAME = "HeavyItemSCPs";
-        const string PLUGIN_VERSION = "1.3.0";
+        const string PLUGIN_VERSION = "1.3.1";
 
         public static Plugin PluginInstance;
         public static ManualLogSource LoggerInstance;
@@ -79,6 +79,8 @@ namespace HeavyItemSCPs
         public static ConfigEntry<float> config178LensDistortion;
         public static ConfigEntry<float> config178ChromaticAberration;
         public static ConfigEntry<string> config178ColorTint;
+        public static ConfigEntry<bool> config178SeeScrapThroughWalls;
+        public static ConfigEntry<float> config178SeeScrapRange;
 
         // SCP-178-1 Configs
         public static ConfigEntry<int> config1781MaxCountInside;
@@ -90,6 +92,9 @@ namespace HeavyItemSCPs
         public static ConfigEntry<float> config1781PostObservationTime;
         public static ConfigEntry<float> config1781DistanceToAddAnger;
         public static ConfigEntry<float> config1781ObservationGracePeriod;
+        public static ConfigEntry<bool> config1781UsePercentageBasedCount;
+        public static ConfigEntry<float> config1781MaxPercentCountInside;
+        public static ConfigEntry<float> config1781MaxPercentCountOutside;
 
         // SCP-323
         public static ConfigEntry<bool> configEnableSCP323;
@@ -191,6 +196,8 @@ namespace HeavyItemSCPs
             config178LensDistortion = Config.Bind("SCP-178 3D Effects", "Lens Distortion", -0.2f, "Changes the lens distortion effect of the 3D glasses.");
             config178ChromaticAberration = Config.Bind("SCP-178 3D Effects", "Chromatic Aberration", 3f, "Changes the chromatic aberration effect of the 3D glasses.");
             config178ColorTint = Config.Bind("SCP-178 3D Effects", "Color Tint", "500,0,500", "Changes the RGB color tint effect of the 3D glasses.");
+            config178SeeScrapThroughWalls = Config.Bind("SCP-178", "See Scrap Through Walls", true, "Changes whether or not the 3D glasses can allow you to see scrap through walls.");
+            config178SeeScrapRange = Config.Bind("SCP-178", "See Scrap Range", 50f, "Changes the range at which the 3D glasses can see scrap.");
 
             // SCP-1781
             config1781MaxCountOutside = Config.Bind("SCP-1781", "Max count outside", 50, "The maximum number of SCP-178-1 instances that can be spawned outside. -1 spawns on all ai nodes.");
@@ -201,7 +208,10 @@ namespace HeavyItemSCPs
             config1781WanderingRadius = Config.Bind("SCP-1781", "Wandering radius", 5f, "The radius around SCP-178-1 spawn position that they will roam around in.");
             config1781WanderingWaitTime = Config.Bind("SCP-1781", "Wandering wait time", 5f, "When spawned, SCP-178-1 will pick a random position in their wandering radius and walk to it. This determines how long they will wait until picking another position to walk to.");
             config1781DistanceToAddAnger = Config.Bind("SCP-1781", "Distance to add anger", 10f, "The distance you need to be from SCP-178-1 to increase their anger meter when looking at them.");
-            config1781ObservationGracePeriod = Config.Bind("SCP-1781", "Observation grace period", 5f, "The time it takes for SCP-178-1 instances to start getting angry after staring at them.");
+            config1781ObservationGracePeriod = Config.Bind("SCP-1781", "Observation grace period", 1f, "The time it takes for SCP-178-1 instances to start getting angry after staring at them.");
+            config1781UsePercentageBasedCount = Config.Bind("SCP-1781 Percentage Based", "Use percentage based count", true, "If true, when putting on the 3D glasses, instead of using max count, it will get the amount of AI nodes and times it by this value to get the amount of SCP-178-1 instances it should spawn.");
+            config1781MaxPercentCountInside = Config.Bind("SCP-1781 Percentage Based", "Max percent count inside", 0.5f, "The percentage of inside AI nodes that should have SCP-178-1 instances spawned on them.");
+            config1781MaxPercentCountOutside = Config.Bind("SCP-1781 Percentage Based", "Max percent count outside", 0.5f, "The percentage of outside AI nodes that should have SCP-178-1 instances spawned on them.");
 
             // SCP-323
             configEnableSCP323 = Config.Bind("SCP-323", "Enable SCP-323", true, "Whether or not SCP-323 can spawn as scrap.");